refactor(wallet): split wallet service into focused modules
